Refactor test and fixtures for User.
